Changed values in map directly to reflect 1.0 pbr, instead of dividing all winnings by 2. Need to check math on how PBR was 2.0 before halving everything

+++ SlotsProto/CreditStateTracker.h
@@ -9,27 +9,27 @@ public:
    
    //Constructors, always needs a starting amount of credits (which is determiend from the configuration)
    CreditStateTracker();
-   CreditStateTracker(size_t, std::string, const std::string&);
+   CreditStateTracker(double, std::string, const std::string&);
 
    ~CreditStateTracker();
 
-   bool gameRequested(const size_t& bet_price);
-   void gameWon(const size_t& credits_won);
+   bool gameRequested(const double& bet_price);
+   void gameWon(const double& credits_won);
    void updateGameStats();
    void printgameStats();
 
    //Get Function for credits
-   int getCredits() const { return credits; }
+   double getCredits() const { return credits; }
 
    //Get functions for recorded numbers
-   size_t getCreditsUsed() const { return credits_used; }
-   size_t getCreditsWon() const { return credits_won; }
+   double getCreditsUsed() const { return credits_used; }
+   double getCreditsWon() const { return credits_won; }
    size_t getRoundsPlayed() const { return rounds_played; }
    size_t getRoundsWon() const { return rounds_won; }
 
    //Get functions for calculated numbers
-   float getPayBackRate() const { return pay_back_rate; }
-   float getHitRate() const { return hit_rate; }
+   double getPayBackRate() const { return pay_back_rate; }
+   double  getHitRate() const { return hit_rate; }
 
    //CreditStateTracker operator=(const CreditStateTracker& newCST);
 private:
@@ -40,17 +40,17 @@ private:
    std::ofstream outputfile;
 
    //Tracked Numbers
-   int credits; // THIS NEEDS TO BE A SIZE_T EVENTUALLY. THIS IS OUR COMPILER WARNING ABOUT SIZE_T TO INT CONVERSION. CHANGE WHEN NOT TESTING.
+   double credits; // THIS NEEDS TO BE A SIZE_T EVENTUALLY. THIS IS OUR COMPILER WARNING ABOUT SIZE_T TO INT CONVERSION. CHANGE WHEN NOT TESTING.
 
    //Recorded Numbers
-   size_t credits_used{ 0 };
-   size_t credits_won{ 0 };
+   double credits_used{ 0 };
+   double credits_won{ 0 };
    size_t rounds_played{ 0 };
    size_t rounds_won{ 0 };
 
    // Calculated Numbers
-   float pay_back_rate{ 0 };
-   float hit_rate{ 0 };
+   double pay_back_rate{ 0 };
+   double hit_rate{ 0 };
 
    void updatePayBackRate();
    void updateHitRate();
